requirements.txt
streamlit
google-genai
transformers
torch
soundfile
librosa
python-dotenv
numpy
audio-recorder-streamlit